Make HeritageList as reusable
